// shopme-client/src/types/carouselTypes.ts
// Types for carousel images based on the Java model:
// public class CarouselImageResponse {
//    private Integer id;
//    private String image;
//    private String name;
//    private String content;
// }

export interface CarouselImageResponse {
  id: number;
  image: string;
  name: string;
  content: string;
}